% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestMap.R
\name{bestMap}
\alias{bestMap}
\title{Title}
\usage{
bestMap(L1, L2)
}
\arguments{
\item{L1}{the true label vector}

\item{L2}{the predicted label vector}
}
\value{
label vector
}
\description{
Use the Hungarian algorithm to optimally adjust the L2 label assignments so that they best match those of L1.
}
