% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringMeasure.R
\name{ClusteringMeasure}
\alias{ClusteringMeasure}
\title{Title}
\usage{
ClusteringMeasure(Y, predY)
}
\arguments{
\item{Y}{the true label vector}

\item{predY}{the predicted label vector}
}
\value{
list containing ACC (accuracy), MIhat (normalized mutual information), and ARI
}
\description{
Title
}
