% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_range.R
\name{get_cluster_range}
\alias{get_cluster_range}
\title{Determine Range of Candidate Cluster Numbers}
\usage{
get_cluster_range(sample_size, custom_range = NULL)
}
\arguments{
\item{sample_size}{An integer specifying the number of samples in the dataset.}

\item{custom_range}{An optional numeric vector of length 2 specifying
the lower and upper bounds of the cluster range.
If provided, this range is returned directly.}
}
\value{
An integer sequence representing the range of candidate cluster numbers.
}
\description{
This function determines an appropriate range of candidate cluster numbers
for clustering analysis, based either on a user-specified range or
automatically depending on the sample size.
}
