% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutualInfo.R
\name{MutualInfo}
\alias{MutualInfo}
\title{Title}
\usage{
MutualInfo(L1, L2)
}
\arguments{
\item{L1}{the first label vector}

\item{L2}{the second label vector}
}
\value{
the normalized mutual information (NMI) value
}
\description{
Title
}
