% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normcol_lessequal.R
\name{normcol_lessequal}
\alias{normcol_lessequal}
\title{Normalize Columns of a Matrix to Unit Length (If Necessary)}
\usage{
normcol_lessequal(matin)
}
\arguments{
\item{matin}{A numeric matrix whose columns are to be normalized.}
}
\value{
A numeric matrix with columns normalized such that
their Euclidean norms are ≤ 1.
}
\description{
This function normalizes each column of the input matrix so that
its Euclidean norm does not exceed 1.
If a column already has a norm less than or equal to 1,
it remains unchanged.
}
