% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateH.R
\name{UpdateH}
\alias{UpdateH}
\title{Title}
\usage{
UpdateH(M, W, L, lambda)
}
\arguments{
\item{M}{data matrix M}

\item{W}{the iteratively updated W matrix}

\item{L}{Laplacian matrix}

\item{lambda}{feature difference penalty parameter}
}
\value{
the updated matrix H
}
\description{
Title
}
