% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateW.R
\name{UpdateW}
\alias{UpdateW}
\title{updated matrix W}
\usage{
UpdateW(Coef, Data, D_Mat)
}
\arguments{
\item{Coef}{latent representation matrix H}

\item{Data}{data matrix M}

\item{D_Mat}{the iteratively updated W matrix}
}
\value{
the updated matrix W
}
\description{
updated matrix W
}
