% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateZ.R
\name{UpdateZ}
\alias{UpdateZ}
\title{Title}
\usage{
UpdateZ(H, G, beta, lambda)
}
\arguments{
\item{H}{a d1 × n matrix (columns represent samples)}

\item{G}{an n × n initial similarity matrix}

\item{beta}{regularization parameter}

\item{lambda}{feature difference penalty parameter}
}
\value{
an n × n similarity matrix Z with non-negative entries and rows summing to 1
}
\description{
Title
}
